curlpw: print a newline after password prompt
Since the password is being input with no echoing to the terminal, it appears to
hang (especially if the curl request takes a while) after pressing enter, even
when the password was successfully read. Inserting a manual echo moves the
cursor to the next line, giving the user feedback that their password input is
complete.